requirements.txt
numpy==1.22.0
Pillow==10.2.0
protobuf==3.18.3
six==1.12.0
tensorboardX==1.8
torch==1.13.1
torchvision==0.14.1   # 已修改为与 torch 对应的版本
tqdm==4.35.0
opencv-python
imageio==2.9.0
matplotlib==3.8.4